% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR1.R
\name{AR1}
\alias{AR1}
\title{Create an AR1 matrix}
\usage{
AR1(N, sigma, rho)
}
\arguments{
\item{sigma}{how much noise exists in the AR1 process}

\item{rho}{how much each obs should be correlated with the obs before}

\item{n}{number of observations in the AR1 process}
}
\value{
Variance-Covrariance matrix of AR1 process
}
\description{
Given values for rho and sigma creates an AR1 matrix
}
\examples{
AR1(20, .95, 3)

}

